fix: move factory-validation API tests to servertests for backend-size
